requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit-ace>=0.1.1
streamlit-extras>=0.3.0
plotly>=5.17.0
pandas>=2.0.0